Move GuideHighlightTour type to exhibition-guides types
